reorganize the project
